Some updates and refactoring
