+++ Task.md
@@ -6,7 +6,6 @@
     4. Админка. Страницы просмотра для Параграфы, Оружие, Заметки, Предметы, Противники.
     5. Заполнить таблицу Параграфы, Оружие, Заметки, Предметы аналогично Противникам
 
-
 Первая итерация
 ------
     1. При нажатии на кнопку "Новая игра" создается новый персонаж.
