+++ changelog/changelog.md
@@ -30,3 +30,5 @@
 - Добавлена текстовая часть главной страницы.
 - Добавлена часть для выбора вариантов на главной.
 - Добавлен блок для изображений и заметок в правом меню главной страницы.
+
+
