+++ src/app/Http/Requests/Core/StoreEnemyRequest.php
