+++ src/app/Http/Requests/Core/StoreNoteRequest.php
