+++ src/app/Http/Requests/Core/StoreWeaponRequest.php
