+++ src/resources/views/core/book/partials/choose_part.blade.php
