+++ src/resources/views/core/book/partials/left_sidebar.blade.php
