+++ src/resources/views/core/book/partials/text_part.blade.php
