+++ src/resources/views/core/book/partials/top_menu.blade.php
