+++ src/resources/views/core/index/partials/content.blade.php
